+++ nuxt.config.js
@@ -38,7 +38,7 @@ export default {
   /*
    ** Plugins to load before mounting the App
    */
-  plugins: ["~/plugins/contentful", "~/plugins/posts","~/plugins/persons"],
+  plugins: ["~/plugins/contentful", "~/plugins/posts","~/plugins/persons", "~/plugins/projects"],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
   components: true,