+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [CalVer Version Scheme](https://calver.org/).
 ### Added
 
 - RSS, using `@nuxtjs/feed` and porting [Dawntraoz's article](https://www.dawntraoz.com/blog/adding-rss-feed-to-my-blog-made-with-nuxt-storyblok) to use Contentful.
+- Projects Contentful content model
 
 ## [v2021.04.0]